Implement injection of a forkserver into the compiler process
Implement a forkserver inspired by AFL fuzzer. Unlike the original
design, this one relies on library preloading and seccomp to
eliminate static or dynamic recompilation of the compiler.

For single-threaded, single-process compilers this can be
several times faster, provided that the interacting
with any input files is delayed until absolutely necessary.

For example, it is the case when the compiler process first
loads the standard library for several seconds and then
processes scratch files in several milliseconds.